requirements.txt
streamlit
PyMuPDF
groq
python-dotenv
